maint: break out lp deps for easy testing
